The proxy functionality has been refactored. The Java 11 HTTP client only supports the HTTP proxy type.
