+++ src/test/java/coresearch/cvurl/io/request/RequestBuilderTest.java
@@ -3,10 +3,11 @@ package coresearch.cvurl.io.request;
 import coresearch.cvurl.io.constant.HttpMethod;
 import coresearch.cvurl.io.internal.configuration.RequestConfiguration;
 import coresearch.cvurl.io.model.CVurlConfig;
+import coresearch.cvurl.io.model.CVurlProxy;
+import coresearch.cvurl.io.request.proxy.CVurlProxySelector;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.net.Proxy;
 import java.net.ProxySelector;
 import java.util.Optional;
 
@@ -15,7 +16,7 @@ import static org.mockito.Mockito.*;
 @SuppressWarnings("rawtypes")
 class RequestBuilderTest {
 
-    private static final CVurlProxy CVURL_PROXY = CVurlProxy.of(Proxy.Type.HTTP, "proxy.host", 90);
+    private static final CVurlProxy CVURL_PROXY = CVurlProxy.of("proxy.host", 90);
     private static final String uri = "http://localhost:8181/";
     private CVurlProxySelector proxySelectorSpy;
     private RequestBuilder builder;
@@ -35,7 +36,7 @@ class RequestBuilderTest {
         when(config.getProxySelector()).thenReturn(Optional.of(proxySelectorSpy));
 
         //when
-        builder.withProxy(CVURL_PROXY);
+        builder.proxy(CVURL_PROXY);
 
         //then
         verifyAddProxyCalled(1);
@@ -47,7 +48,7 @@ class RequestBuilderTest {
         when(config.getProxySelector()).thenReturn(Optional.empty());
 
         //when
-        builder.withProxy(CVURL_PROXY);
+        builder.proxy(CVURL_PROXY);
 
         //then
         verifyAddProxyCalled(0);
@@ -59,7 +60,7 @@ class RequestBuilderTest {
         when(config.getProxySelector()).thenReturn(Optional.of(ProxySelector.getDefault()));
 
         //when
-        builder.withProxy(CVURL_PROXY);
+        builder.proxy(CVURL_PROXY);
 
         //then
         verifyAddProxyCalled(0);
